source/docs/003_Function_Calling的工作流程/my-weather-function-calling/pyproject.toml
[project]
name = "my-weather-function-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
